feat: Add support for changing SU upload+response strategies
This commit adds support for 3 operating modes for SU's that trade-off speed of response with safety.

`aggressive`: The SU does not write the data to any storage media before returning the attestation to the caller.
`local_confirmation`: The SU returns the attestation only after ensuring that the data item is stored on their local drive.
`remote_confirmation`: Only return after the data has been sent to the SU's bundling service.