wip: implement cache-lookup and execution de-dupe in hb_converge
This commit moves much of the functionality previously in `hb_process` into the core of `hb_converge`. By structuring things this way we enable any execution -- not just processes -- to have a persistent in-memory cache.

Much of the multi-process orchestration logic could be moved into a separate module if we prefer. At the moment the core Converge execution pipeline is ~250 lines.