fix: encode functions as binaries in structured@1.0.
This is not reversible, but is necessary in order to avoid the keys being skipped completely. Skipped keys function lead to a `device` attribute that is an empty map from the POV of the network. This in turn created a situation in which `hb_cache` would attempt to write a message whose `device` was empty -- containing another empty map device etc., recursively