draft: support IDs in header names for httpsig@1.0 via percent encoding
This proposal implements percent encoding and decoding for IDs that are passed over the wire as HTTP headers. This generates rather long and ugly header names, but critically circumvents the issue of HTTP/2 and HTTP/3 requiring strictly lower-case keys during transmission, and Arweave's usage of mixed case keys. Notably, this only applies to header names -- not their associated values.

One improvement to this patch to minimize the  jank could be to ensure that multipart message elements do not have this encoding applied. Their headers, however, would still require it.