chore: rename `attestation` to `commitment`.
This reflects that a cryptographic proof in AO can be a tradition attestation/signature, but it can also be a variety of different forms of 'commitment', which may not have an explicit 'attestor'.

Consider, for example, that `hmac-sha256` is a form of 'signature' in the HTTP Signed Messages specification, but it does not pertain inherently to a 'signer'. Additionally, when ZK proofs are implemented for hashpath resolutions, they too should utilize the `attestation` mechanisms. `Commitment` much more accurately describes these primitives.