impr: Recursive device inheritance
This commit enables the use of inheritance of key resolvers recursively between devices. A device may now specify the `info/default` key in the form of a device name or module which the AO-Core engine will use to resolve keys that do not match in the primary device. AO-Core already performed a similar operation while defaulting from a given device to `message@1.0` when a key did not match, while this commit opens up that system to allow for standard OOP-style inheritance structures.