impr: tidy noisy event groups and improve prometheus management
Removes some particularly spammy event groups from prometheus counts, such that excessive logging does not cause issues on busy nodes.

Additionally, if the memory of the prometheus logging server exceeds a standardized bound (1 GiB for now), the Erlang process is killed and rebooted by the next caller. This loses the logs therein, but is preferable over causing the server to significantly disfunction.